separate relationshipIndicators js
